Roll back top. sort fix regarding HelperBlock/calibration temporarily, and update all "broyden" to "broyden_custom" in line with renaming of the solver
Reverting the fix of checking if "isinstance(block, HelperBlock)" within the construction of the output graph, since this fix was not fully correct. Currently under this commit tests now pass, so need to implement a proper fix of this that passes tests and accounts for the fact that calibration sometimes may contain initial values of variables/parameters, and not just fixed values.