Allow two-argument unknown initial values (lower and upper bounds) for multivariate solver, notes below.
This was implemented specifically to allow symmetry of calling solve_steady_state() with and without helper blocks, where in the case of Krusell Smith, we need to use a multivariate solver when not using helper blocks, but we need to use a univariate solver when using helper blocks. The multivariate solver requires initial values, whereas the univariate solver requires bounds. To impose a uniform interface, we take the midpoint as the initial value for the multivariate solver when bounds are provided.