Fix bug with two_asset helper block implementation. Notes below:
- Previously the vphi/wnkpc unknown/target pair were treated as non-excludable because the helper block they show up in, partial_ss_step2, had a non-helper block
as a dependency or in other words depended on computed outputs from the main DAG.
This rule of thumb helped us to separate the notion of a helper block that did not depend on computed outputs from the main DAG, i.e. partial_ss_step_1,
and those that did depend on computed outputs from the main DAG. The issue was that the root-finding algorithm (broyden_custom) tried to form a Jacobian with
respect to the unknowns/targets of the main DAG to update its guess for the unknowns, which included the unknown/target pair vphi/wnkpc. However,
wnkpc was written in a way such that for *any* value of vphi it would be equal to 0. and hence the Jacobian was singular in that dimension.
It is sensible that it was written this way since even though helper blocks may appear later on in the DAG/may be functions of computed outputs from the main DAG they
still represent analytical solutions for an unknown given a target and that computed output.
Hence, I have broadened the exclusion behavior of the _solve_for_unknowns_w_helper_blocks to treat all helper unknowns/targets as "excluded" in the strict sense that
one should not have to *directly* numerically solve for them as with other unknowns/targets, but rather they may still need to be computed *indirectly* from other
computed outputs.